cleanup and add docs
